Serial/requirements.txt
pyserial==3.5b0
tqdm==4.66.1
